requirements.txt
streamlit
boto3
openai
google-genai
ipython
streamlit_chat
deepeval
